requirements.txt
mealie-client
python-dotenv
Wikidata
